refactor: rename command for cluster
