+++ src/EntityDb.MongoDb.Provisioner/Program.cs
@@ -24,7 +24,7 @@ internal static class Program
 
         CreateRole.AddTo(rootCommand);
         CreateUser.AddTo(rootCommand);
-        CreateCollections.AddTo(rootCommand);
+        CreateCollectionsCluster.AddTo(rootCommand);
         CreateCollectionsDirect.AddTo(rootCommand);
 
         return rootCommand.InvokeAsync(args);
